+++ src/components/index.js
@@ -1,0 +1,3 @@
+import GNB from './GNB/GNB';
+
+export { GNB };
