+++ src/pages/index.js
@@ -1,3 +1,4 @@
 import Login from './Login/Login';
+import Main from './Main/Main';
 
-export { Login };
+export { Login, Main };
